+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * support for `gcp` backend
 * add node exporter to the prometheus exporter setup
 * if feature file is not provided and cluster size is larger than 1, bail with critical error
+* add `legacy` option to files `upload` and `download` functions
 
 #### 5.4.6
 * update all libraries (dependencies)